third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement.getBBox-05-expected.txt

This is a testharness.js-based test.
[FAIL] image with fill, stroke, markers and clipping
  assert_approx_equals: image1.getBBox().x {"fill":true,"stroke":true,"markers":true,"clipped":true} expected 160 +/- 0.1 but got 150
[FAIL] path with fill, stroke, markers and clipping
  assert_approx_equals: path1.getBBox().x {"fill":true,"stroke":true,"markers":true,"clipped":true} expected 6.2 +/- 0.1 but got 10
[FAIL] path with fill and clipping
  assert_approx_equals: path4.getBBox().width {"fill":true,"stroke":false,"markers":false,"clipped":true} expected 90 +/- 0.1 but got 100
[FAIL] path with stroke
  assert_approx_equals: path1.getBBox().x {"fill":false,"stroke":true,"markers":false,"clipped":false} expected 6.2 +/- 0.1 but got 10
[FAIL] path with markers
  assert_approx_equals: path1.getBBox().width {"fill":false,"stroke":false,"markers":true,"clipped":false} expected 112 +/- 0.1 but got 100
[FAIL] use with fill, stroke, markers and clipping
  assert_approx_equals: use1.getBBox().y {"fill":true,"stroke":true,"markers":true,"clipped":true} expected 66 +/- 0.1 but got 70
[FAIL] foreignObject with fill, stroke, markers and clipping
  assert_approx_equals: fo1.getBBox().x {"fill":true,"stroke":true,"markers":true,"clipped":true} expected 53 +/- 0.1 but got 2
[FAIL] masking-path-07-b.html with fill, stroke, markers and clipping
  assert_approx_equals: rect-1.getBBox().x {"fill":true,"stroke":true,"markers":true,"clipped":true} expected 10 +/- 0.1 but got 0
Harness: the test ran to completion.
